Update the model

+++ src/main/java/ru/internetprovider/model/Dao.java
@@ -2,8 +2,8 @@ package ru.internetprovider.model;
 
 public interface Dao<T> {
 
-    T get(long id);
-    void delete(long id);
+    T get(int id);
+    void delete(int id);
 
 
 }
